+++ CHANGELOG.md
@@ -4,7 +4,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.31.0] - 2020-06-01
 ### Added
+- VIMEET-2124 - Share a video in a webinar
+- VIMEET-2139 - Do not hide participants stream; maximize screen or video sharing
+
+## [2.30.0] - 2020-05-29
+### Fixed
+- MV-237 - Check user imported and no password in register path
+
+## [2.29.0] - 2020-05-25
+### Added
+- VIMEET-2117 - Webinar audio/video settings
 - VIMEET-2088 - Sort sheet participants
 - VIMEET-2101 - Add end sound, image and message to visio meeting
 - VIMEET-2105 - Interactive webinar